requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
jinja2==3.1.4
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
langchain-core==1.0.4
langchain-google-genai==3.0.3

